refactor: use subpath imports for types module reference
