+++ packages/vite/package.json
@@ -42,7 +42,9 @@
     "#module-sync-enabled": {
       "module-sync": "./misc/true.js",
       "default": "./misc/false.js"
-    }
+    },
+    "#types/*": "./types/*.d.ts",
+    "#dep-types/*": "./src/types/*.d.ts"
   },
   "files": [
     "bin",
@@ -113,7 +115,6 @@
     "cors": "^2.8.5",
     "cross-spawn": "^7.0.6",
     "debug": "^4.4.3",
-    "dep-types": "link:./src/types",
     "dotenv": "^17.2.3",
     "dotenv-expand": "^12.0.3",
     "es-module-lexer": "^1.7.0",
@@ -147,7 +148,6 @@
     "strip-literal": "^3.1.0",
     "terser": "^5.44.0",
     "tsconfck": "^3.1.6",
-    "types": "link:./types",
     "ufo": "^1.6.1",
     "ws": "^8.18.3"
   },
