+++ packages/vite/src/node/plugins/importMetaGlob.ts
@@ -13,9 +13,9 @@ import type {
 import type { CustomPluginOptions, RollupAstNode, RollupError } from 'rollup'
 import MagicString from 'magic-string'
 import { stringifyQuery } from 'ufo'
-import type { GeneralImportGlobOptions } from 'types/importGlob'
 import { parseAstAsync } from 'rollup/parseAst'
 import { escapePath, glob } from 'tinyglobby'
+import type { GeneralImportGlobOptions } from '#types/importGlob'
 import type { Plugin } from '../plugin'
 import type { EnvironmentModuleNode } from '../server/moduleGraph'
 import type { ResolvedConfig } from '../config'
