+++ packages/vite/src/node/server/hmr.ts
@@ -2,8 +2,8 @@ import fsp from 'node:fs/promises'
 import path from 'node:path'
 import { EventEmitter } from 'node:events'
 import colors from 'picocolors'
-import type { CustomPayload, HotPayload, Update } from 'types/hmrPayload'
 import type { RollupError } from 'rollup'
+import type { CustomPayload, HotPayload, Update } from '#types/hmrPayload'
 import type {
   InvokeMethods,
   InvokeResponseData,
