+++ packages/vite/src/node/server/middlewares/error.ts
@@ -2,8 +2,8 @@ import path from 'node:path'
 import { stripVTControlCharacters as strip } from 'node:util'
 import colors from 'picocolors'
 import type { RollupError } from 'rollup'
-import type { Connect } from 'dep-types/connect'
-import type { ErrorPayload } from 'types/hmrPayload'
+import type { Connect } from '#dep-types/connect'
+import type { ErrorPayload } from '#types/hmrPayload'
 import { pad } from '../../utils'
 import type { ViteDevServer } from '../..'
 import { CLIENT_PUBLIC_PATH } from '../../constants'
