+++ packages/vite/src/node/server/middlewares/htmlFallback.ts
@@ -1,6 +1,6 @@
 import path from 'node:path'
 import fs from 'node:fs'
-import type { Connect } from 'dep-types/connect'
+import type { Connect } from '#dep-types/connect'
 import { createDebugger } from '../../utils'
 import { cleanUrl } from '../../../shared/utils'
