+++ packages/vite/src/node/server/middlewares/proxy.ts
@@ -1,7 +1,7 @@
 import type * as http from 'node:http'
 import * as httpProxy from 'http-proxy-3'
-import type { Connect } from 'dep-types/connect'
 import colors from 'picocolors'
+import type { Connect } from '#dep-types/connect'
 import { createDebugger } from '../../utils'
 import type { CommonServerOptions, ResolvedConfig } from '../..'
 import type { HttpServer } from '..'
