+++ packages/vite/src/node/watch.ts
@@ -1,9 +1,9 @@
 import { EventEmitter } from 'node:events'
 import path from 'node:path'
-import type { FSWatcher, WatchOptions } from 'dep-types/chokidar'
 import type { OutputOptions } from 'rollup'
 import colors from 'picocolors'
 import { escapePath } from 'tinyglobby'
+import type { FSWatcher, WatchOptions } from '#dep-types/chokidar'
 import { withTrailingSlash } from '../shared/utils'
 import { arraify, normalizePath } from './utils'
 import type { Logger } from './logger'
